requirements.txt
alembic==1.7.6
APScheduler==3.8.1
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.24.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==7.1.2
CMRESHandler==1.0.0
elasticsearch==7.17.0
fastapi==0.65.1
greenlet==1.1.2
h11==0.9.0
httptools==0.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.8.6
pydantic==1.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==3.12
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.3
starlette==0.14.2
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.11.8
uvloop==0.16.0
websockets==8.1
zipp==3.7.0
